fix: Refactor LUIS, QNA and Speech to use a single base component (#6923)
* Refactor LUIS, QNA and Speech to use a single base component

* fix height, sort order

* fix typing

* use shared colors

* refactor handoff instructions to share some text

* Update BotProjectsSettingsTabView.tsx

* refactor qna maker provisioning into a background process

* fix linter quible:w

* remove dep

* rename setVisible to onToggleVisibility

* Use shared color

* add conditional around missing hostname

* add conditional around missing hostname

* remove duplicate import
